src: clean-cut move to release/11.0.0
Taken from: FreeBSD
Commit ref: d508cb8c8f

+++ contrib/llvm/tools/lldb/source/Commands/CommandObjectQuit.cpp
@@ -7,8 +7,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "lldb/lldb-python.h"
-
 #include "CommandObjectQuit.h"
 
 // C Includes
@@ -17,6 +15,7 @@
 // Project includes
 #include "lldb/Interpreter/CommandInterpreter.h"
 #include "lldb/Interpreter/CommandReturnObject.h"
+#include "lldb/Target/Process.h"
 
 using namespace lldb;
 using namespace lldb_private;
@@ -53,7 +52,7 @@ CommandObjectQuit::ShouldAskForConfirmation (bool& is_a_detach)
             continue;
         const TargetList& target_list(debugger_sp->GetTargetList());
         for (uint32_t target_idx = 0;
-             target_idx < target_list.GetNumTargets();
+             target_idx < static_cast<uint32_t>(target_list.GetNumTargets());
              target_idx++)
         {
             TargetSP target_sp(target_list.GetTargetAtIndex(target_idx));
